+++ src/plugins/image/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { DocRenderer } from "../../types";
+import { DocRenderer } from "../..";
 
 const ImageProxyRenderer: DocRenderer = (props) => {
   const {
