+++ src/plugins/msg/index.tsx
@@ -2,7 +2,7 @@
 import React, { FC, useEffect, useState } from "react";
 import styled from "styled-components";
 import { MSGErrorResult, MSGFileData, MSGReader } from "wl-msg-reader";
-import { DocRenderer, IStyledProps } from "../../types";
+import { DocRenderer, IStyledProps } from "../..";
 import { arrayBufferFileLoader } from "../../utils/fileLoaders";
 
 const MSGRenderer: DocRenderer = ({ mainState: { currentDocument } }) => {
