+++ src/plugins/pdf/components/PDFPagination.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useContext } from "react";
 import styled from "styled-components";
 import { Button } from "../../../components/common";
-import { IStyledProps } from "../../../types";
+import { IStyledProps } from "../../..";
 import { PDFContext } from "../state";
 import { setCurrentPage } from "../state/actions";
 import { NextPDFNavIcon, PrevPDFNavIcon } from "./icons";
