+++ src/plugins/pdf/components/pages/PDFSinglePage.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useContext } from "react";
 import { Page } from "react-pdf";
 import styled from "styled-components";
-import { IStyledProps } from "../../../../types";
+import { IStyledProps } from "../../../..";
 import { PDFContext } from "../../state";
 
 interface Props {
